+++ multivar_horner/global_settings.py
@@ -9,6 +9,7 @@ ID_ADD = True  # ATTENTION: is being used in helpers_fcts_numba.py/eval_recipe()
 INT_DTYPE = numpy.int64  # i8 =  8byte integer
 UINT_DTYPE = numpy.uint32  # u4 =  4byte unsigned integer
 FLOAT_DTYPE = numpy.float64  # f8 =  8byte float
+BOOL_DTYPE = numpy.bool
 
 # python typing
 TYPE_1D_FLOAT = Union[numpy.ndarray, List[float]]